+++ src/Router/Router.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import ErrorPage from '../pages/ErrorPage'
-import PokedexPage from '../pages/PokedexPage/PokedexPage'
-import PokemonDetailPage from '../pages/PokemonDetailPage/PokemonDetailPage'
-import PokemonsListPage from '../pages/PokemonsListPage/PokemonsListPage'
+import PokedexPage from '../pages/PokedexPage'
+import PokemonDetailPage from '../pages/PokemonDetailPage'
+import PokemonsListPage from '../pages/PokemonsListPage'
 
 const Router = () => {
   return (
@@ -11,7 +11,7 @@ const Router = () => {
         <Routes>
             <Route path='/' element= {<PokemonsListPage/>} />
             <Route path='/pokedex' element={<PokedexPage/>} />
-            <Route path='/pokedex/id' element={<PokemonDetailPage/>}/>
+            <Route path='/pokemon/detalhes/:namePokemon' element={<PokemonDetailPage/>}/>
             <Route path='*' element={<ErrorPage/>}/>
         </Routes>
     </BrowserRouter>
