+++ src/Router/coordinator.js
@@ -6,6 +6,6 @@ export const goToPokedexPage = (navigate) =>{
     navigate('/pokedex')
 }
 
-export const goToPokemonDetailPage = (navigate, id)=>{
-    navigate('/pokedex/id')
+export const goToPokemonDetailPage = (navigate, namePokemon)=>{
+    navigate(`/pokemon/detalhes/${namePokemon}`)
 }
