+++ src/pages/PokedexPage.js
@@ -1,19 +1,23 @@
+import {
+  Box,
+  Grid,
+  Heading
+}
+  from '@chakra-ui/react'
 import React, { useContext } from 'react'
-import { Box, Grid, Heading } from '@chakra-ui/react'
-import Header from '../../components/Header/Header'
-import { GlobalContext } from '../../contexts/GlobalStateContext'
-import PokemonCard from '../../components/PokemonCard/PokemonCard'
+import Header from '../components/Header'
+import { GlobalContext } from '../contexts/GlobalStateContext'
 
 const PokedexPage = () => {
   const context = useContext(GlobalContext)
-  const { pokedex, removePokedex } = context
+  const { renderPokedex } = context
 
   return (
     <>
       <Header />
       <Box
         w='100%'
-        h={1000}
+        h={2000}
         bg={'gray.600'}
         padding="40px 32px"
       >
@@ -32,14 +36,7 @@ const PokedexPage = () => {
           columnGap={4}
           paddingTop='50px'
         >
-          {pokedex.map((pokemon) => {
-            return <PokemonCard
-              key={pokemon.name}
-              pokemonName={pokemon.name[0].toUpperCase() + pokemon.name.substring(1)}
-              pokemon={pokemon.name}
-              removePokedex={removePokedex}
-            />
-          })}
+          {renderPokedex}
         </Grid>
       </Box>
     </>
