+++ src/pages/PokemonsListPage.js
@@ -1,12 +1,16 @@
-import { Box, Grid, Heading } from '@chakra-ui/react'
+import { 
+  Box, 
+  Grid, 
+  Heading 
+} 
+  from '@chakra-ui/react'
 import React, { useContext } from 'react'
-import Header from '../../components/Header/Header'
-import PokemonCard from '../../components/PokemonCard/PokemonCard'
-import { GlobalContext } from '../../contexts/GlobalStateContext'
+import Header from '../components/Header'
+import { GlobalContext } from '../contexts/GlobalStateContext'
 
 const PokemonsListPage = () => {
   const context = useContext(GlobalContext)
-  const { pokemons, addToPokedex } = context
+  const { renderPokemonList } = context
 
   return (
     <>
@@ -16,7 +20,7 @@ const PokemonsListPage = () => {
         h='100%'
         bg={'gray.600'}
         padding="40px 32px"
-        >
+      >
         <Heading
           fontFamily="'Poppins', sans-serif"
           fontWeight="700"
@@ -31,14 +35,8 @@ const PokemonsListPage = () => {
           rowGap={10}
           columnGap={4}
           paddingTop='50px'
-          >
-          {pokemons.map((pokemon) => {
-            return <PokemonCard
-              key={pokemon.url}
-              pokemonName={pokemon.name[0].toUpperCase() + pokemon.name.substring(1)}
-              pokemon={pokemon.name}
-              addToPokedex={addToPokedex} />
-          })}
+        >
+          {renderPokemonList}
         </Grid>
       </Box>
     </>
